+++ packages/perseus/src/prompt-types.ts
@@ -17,7 +17,7 @@ import type {NumericInputPromptJSON} from "./widget-ai-utils/numeric-input/promp
 import type {OrdererPromptJSON} from "./widget-ai-utils/orderer/orderer-ai-utils";
 import type {PassagePromptJSON} from "./widget-ai-utils/passage/passage-ai-utils";
 import type {PassageRefPromptJSON} from "./widget-ai-utils/passage-ref/passage-ref-ai-utils";
-import type {RadioPromptJSON} from "./widget-ai-utils/radio/prompt-utils";
+import type {RadioPromptJSON} from "./widget-ai-utils/radio/radio-ai-utils";
 import type {SorterPromptJSON} from "./widget-ai-utils/sorter/prompt-utils";
 import type {UnsupportedWidgetPromptJSON} from "./widget-ai-utils/unsupported-widget";
